requirements.txt
django
gunicorn
graphene_django
django-heroku
django-graphql-jwt
pyjwt==1.7.0